fix: interact_contract_backend
