+++ Backend/src/abi/FundToken.json
@@ -325,4 +325,4 @@
       "stateMutability": "nonpayable",
       "type": "function"
     }
-]
+  ]
