--- src/services/index.ts
@@ -1,2 +1,0 @@
-export * from "./BlockchainService";
-export * from "./FundService";
